backend/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "langchat"
version = "0.0.1"
dependencies = [
  "python-dotenv",
  "langchain",
  "langchain-anthropic",
  "fastapi",
  "uvicorn[standard]",
  "e2b-code-interpreter"
]
